+++ qa/lua/run.lua
@@ -49,6 +49,7 @@ local suites = {
     "test_mod_tweaker_numeric_editor",
     "test_mp_dailies",
     "test_mp_quest_boundary",
+    "test_mp_shilling_ui_policy",
     "test_wt_passive_charge",
     "test_wt_longbow_zoom_probe",
     "test_wt_cwv_ownership",